# [13.0][FIX] resource_booking: performance with recurrent events