`rrule` isn't very performant on iterations. This is used on calendar
events to compute the virtual events set.

We want to avoid calling it too much.

Real case with website_sale_resource_booking (thousands of virtual events):

before: /shop/booking/<id>/schedule: ~25 seconds
after: /shop/booking/<id>/schedule: ~6 seconds

before: /shop/booking/<id>/confirm: ~20 seconds
after: /shop/booking/<id>/confirm: ~3 seconds

before: /shop/checkout: ~18 seconds
after: /shop/checkout: ~2 seconds

before: /shop/address: ~18 seconds
after: /shop/address: ~2 seconds

There's room for improvement anyway

An speed scope for /shop/address before and after:

before:
![image](https://user-images.githubusercontent.com/5040182/172610148-323689c1-b406-4e53-b71f-9a3cdb29f4cc.png)

after:
![image](https://user-images.githubusercontent.com/5040182/172610194-1091f8bd-ad6f-43f0-af01-56815debd3ac.png)




cc @Tecnativa TT37089

ping @pedrobaeza @sergio-teruel @carlosdauden