@yajo 2022-06-09
Have you tested with https://github.com/odoo/odoo/pull/69822?