@chienandalu 2022-06-09
> Have you tested with [odoo/odoo#69822](https://github.com/odoo/odoo/pull/69822)?

Thanks @yajo Yes, I checked it but I wanted to avoid lateral issues from such a big change. And the performance gain isn't that much added to this one, I'm afraid (although there's gain). We'll try to get along with these performance times until the next version and keep recurrent events as low as possible.